refactor: split dashboard into reusable components
